+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # 3.0.5
+**This version includes Breaking Changes**
+- Made EnkaUser#uid, EnkaUser#verificationCode, EnkaUser#verificationExpires, and EnkaUser#verificationCodeRetries nullable.
+- Made Material#stars nullable.
 - Added CachedAssetsManager#_validateCache for cache validation.
 # 3.0.4
 - Made skillDepotId optional on EnkaClient#getCharacterById ([#33](https://github.com/yuko1101/enka-network-api/pull/33))